Move codec types to internal-non-workflow-common
